+++ src/main/java/net/dasik/social/api/group/FlockType.java
@@ -1,0 +1,6 @@
+package net.dasik.social.api.group;
+
+public enum FlockType {
+    AERIAL,
+    TERRESTRIAL
+}